Update merge code
