+++ src/App.css
@@ -20,7 +20,7 @@ a{
 
 
   /* overflow-y: hidden; */
-}
+
 
 /* ============================================= */
 
@@ -41,6 +41,7 @@ a{
   position: relative;
   display: flex;
   align-items: center;
+  
 
   /* background-color: red; */
 }
@@ -127,7 +128,7 @@ a{
   color: #97bf0f;
   background-color: rgb(0, 255, 64);
   border: 3px solid #97bf0f;
-  transform: scale(1.2));
+  transform: scale(1.2);
   transition: all 0.5s ease-in-out;
 }
 
@@ -170,9 +171,10 @@ body {
 
 .side_menu {
   border-right: 1px solid black;  
-  /* position: fixed; */
-  /* max-width: 250px; */
+   position: fixed; 
+  max-width: 250px; 
   background-color:#F8F9FA;
+ 
 }
 .fas {
   cursor: pointer;
