+++ src/Carousel.css
@@ -97,7 +97,7 @@
   color: #97bf0f;
   background-color: rgb(0, 255, 64);
   border: 3px solid #97bf0f;
-  transform: scale(1.2));
+  transform: scale(1.2);
   transition: all 0.5s ease-in-out;
 }
