+++ src/components/doctors/DoctorForm.jsx
@@ -1,8 +1,9 @@
 import { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { addDoctor } from '../../redux/doctors/doctorsSlice';
-import SideMenu from '../SideMenu';
+
 import { useNavigate } from 'react-router-dom';
+import SideMenu from '../SideMenu';
 
 export default function BookForm() {
   const dispatch = useDispatch();
@@ -24,7 +25,7 @@ export default function BookForm() {
       user_id: userId,
     };
     dispatch(addDoctor(newDoctor));
-
+console.log(newDoctor)
     // Reset input fields after adding the doctor
     setName('');
     setSpecialization('');
@@ -37,7 +38,7 @@ export default function BookForm() {
   return (
     <div className='d-flex vh-100 p-0 m-0 '  >
 
-      <Side_menu />
+      <SideMenu/>
 
       <div className='d-flex flex-column align-items-center justify-content-center gap-5  w-100 bg-light' >
